# perf(NODE-6246): Significantly improve memory usage and performance of ObjectId