### Description

This MR refactors ObjectId class to persist strings instead of a raw Buffer, which **greatly** reduces memory usage and improves performance of ObjectIds.

### Why?
- Buffer (or TypedArray) is inefficient at storing small amount of data 
- Buffer of size 12 takes **96 bytes** in v8! ObjectId with buffer is **128 bytes** each
- string of size 24 takes **40 bytes**. ObjectId with String is **72 bytes**

### Performance Improvements

| Operation            | ObjectId (ops/sec) | ObjectId STRING (ops/sec) | Relative Improvement (%) |
|----------------------|----------------|-------------------|--------------------------|
| new ObjectId()       | 10,398,719     | 11,263,765         | **8.32%**                  |
| new ObjectId(string) | 4,965,933      | 14,869,618        | **199.43%**                  |
| deserialize          | 6,955,399      | 6,431,123        | **-7.54%**                  |
| serialize            | 3,881,206     |  3,247,690        | **-16.32%**                  |
| fromExtendedJSON          | 4,821,665      | 50,886,088        | **955.36%**                  |
| toExtendedJSON            | 10,988,941     | 134,783,499        | **1126.54%**                  |
| toHexString          | 10,899,295     | 131,767,192        | **616.63%**                  |
| .equals(objectId)     | 39,653,912     | 44,413,312        | **12.00%**                   |
| .equals(string)   | 7,691,853 | 55,222,764 | **617.94%** |
| .isValid(string) | 137,247 | 21,470,293 | **15543.49%** |


### Real world use cases

In the benchmarks above you can see there are some performance regressions with the raw serialize/deserializing BSON. This is because we now need to convert buffer to/from string instead of just copying bytes, but this regression seems to be moot during "real world" benchmarks. 

I believe the 2 most common use cases are 
- 1. Retrieving documents from the DB and returning them to the user. 
- 2. Creating a new document and persisting to the DB

Using the [MFlix Sample Dataset](https://www.mongodb.com/docs/atlas/sample-data/sample-mflix/) "Comments" documents.

| Operation            | ObjectId (ops/sec) | ObjectId STRING (ops/sec) | Relative Improvement (%) |
|----------------------|----------------|-------------------|--------------------------|
| GET 100 docs (deserialize + stringily)       | 3,762     | 3,855         | **2.46%**                  |
| Create new doc (create + serialize) | 673,069      | 710,472        | **5.56%**                  |

### Memory Improvements

With this change, retained memory usage is reduced by **~45%** (ObjectId size decreases from 128 bytes to 72 bytes). I've also removed an extra buffer allocation during deserialization which reduces the strain on the buffer pool, reducing the likelihood of the internal buffer pool having to allocate another block of `Buffer.poolSize`. 

By reducing the amount of memory that needs to be allocated we are further improving performance since garbage collection is quite expensive.

![image](https://github.com/mongodb/js-bson/assets/2258445/a353b6bb-40d8-472a-abb0-2032138646ea)

#### Example
Using the [MFlix Sample Dataset](https://www.mongodb.com/docs/atlas/sample-data/sample-mflix/) "Comments" collection. Grabbing 1 million documents

```ts
const bef = process.memoryUsage().rss;
const docs = await comments.find({}, { limit: 1000000 }).toArray() // lets grab 1 million documents
const aft = process.memoryUsage().rss;
console.log(`Memory used: ${Math.floor((aft - bef) / 1024 / 1024)}MB`);
```

**BEFORE: 654 MB used**

**AFTER: 343 MB used**

**-47% reduction in memory usage** 

#### cacheBuffer Option

Similar to the previous `cacheHexString` static variable, this MR adds `cacheBuffer` option that also persists the `Buffer` on the ObjectId to speed up some operations that require buffer, such as `.id`

Previous ObjectID w/ cacheHexString VS New ObjectId w/ cacheBuffer

| Operation               | bson (ops/sec) | bsonSTR (ops/sec) | Relative Improvement (%) |
|-------------------------|----------------|-------------------|--------------------------|
| new ObjectId()          | 5,535,118      | 5,481,568         | -0.97%                   |
| new ObjectId() + serialize | 5,621,583    | 5,635,728         | 0.25%                    |
| new ObjectId(string)    | 3,794,209      | 5,363,446         | 41.36%                   |
| deserialize             | 3,788,455      | 6,131,230         | 61.85%                   |
| serialize               | 67,014,370     | 80,864,967        | 20.65%                   |
| toHexString             | 71,372,677     | 79,949,268        | 12.02%                   |
| equals                  | 40,714,896     | 45,167,881        | 10.95%                   |
| equals-string           | 20,697,929     | 42,426,021        | 105.00%                  |
| getTimestamp            | 10,554,392     | 10,591,691        | 0.35%                    |
| createFromTime          | 6,479,506      | 5,801,031         | -10.47%                  |
| isValid                 | 128,998        | 20,637,794        | 15895.85%                |


#### What is changing?

##### Is there new documentation needed for these changes?

This MR deprecates `cacheHexString` which may need to be documented. 

This MR adds `cacheBuffer` which may need to be documented

#### What is the motivation for this change?

We are been running into memory issues while pulling large Mongo result sets into memory, and after lots of memory profiling the issue seems to be related to ObjectId, specifically how memory inefficient Buffer/ArrayBuffer is when storing lots of small Buffers.

We were expecting an ObjectId to consume ~12bytes of memory (+ some overhead), but in reality this consumes 128 bytes per ObjectId (96 bytes for just the Buffer).

I opened an issue with the NodeJS Performance team but this appears to be working as designed: https://github.com/nodejs/performance/issues/173

Storing a string in Node/V8 is much more memory efficient since it's a primitive. A 24 character hex string only consumes 40 bytes of memory, AND it's much faster to serialize/deserialize.

### Release Highlight

<!-- RELEASE_HIGHLIGHT_START -->

### Fill in title or leave empty for no highlight

<!-- RELEASE_HIGHLIGHT_END -->

### Double check the following

- [x] Ran `npm run check:lint` script
- [x] Self-review completed using the [steps outlined here](https://github.com/mongodb/node-mongodb-native/blob/HEAD/CONTRIBUTING.md#reviewer-guidelines)
- [x] PR title follows the [correct format](https://www.conventionalcommits.org/en/v1.0.0/): `type(NODE-xxxx)[!]: description`
  - Example: `feat(NODE-1234)!: rewriting everything in coffeescript`
- [x] Changes are covered by tests
- [x] New TODOs have a related JIRA ticket