@SeanReece 2024-07-04
@H4ad this is great! I was hesitant to touch `generate` but I like what you’ve done here. I’ll try to roll it into this change tomorrow and update the benchmarks. You also got me thinking we could use something like `parseInt(__id.substring(0,8), 16)` to get the time out of the hex string instead of using buffer. If that works and is more performant then we could drop the `cacheBuffer` option entirely. 🤔