@H4ad 2024-07-04
> You also got me thinking we could use something like parseInt(__id.substring(0,8), 16) to get the time out of the hex string instead of using buffer.

Yeah, this will work `new Date(parseInt(this.__id.substring(0,8), 16) * 1000)`.

> If that works and is more performant then we could drop the cacheBuffer option entirely. 

As a compatibility layer, I think keeping the buffer is still worthy to have, in case there is someone relying in something that can only be done via buffer.